% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{calculate_strand_bias_pval_vaf_and_dp_parallel}
\alias{calculate_strand_bias_pval_vaf_and_dp_parallel}
\title{calculate_strand_bias_pval_vaf_and_dp_parallel}
\usage{
calculate_strand_bias_pval_vaf_and_dp_parallel(
  collapsed_read_counts,
  master_indel_record_table_no_dup_reads,
  number_cores,
  mapq_threshold,
  bamPath
)
}
\arguments{
\item{collapsed_read_counts}{collapsed_read_counts}

\item{master_indel_record_table_no_dup_reads}{master_indel_record_table_no_dup_reads}

\item{number_cores}{number_cores}

\item{mapq_threshold}{mapq_threshold}

\item{bamPath}{bamPath}
}
\value{
collapsed_read_counts_with_strand_correction_pvalue_for_writing
}
\description{
calculate_strand_bias_pval_vaf_and_dp_parallel
}
