% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{calculate_strand_bias_pval_vaf_and_dp_parallel_each_indel}
\alias{calculate_strand_bias_pval_vaf_and_dp_parallel_each_indel}
\title{calculate_strand_bias_pval_vaf_and_dp_parallel_each_indel}
\usage{
calculate_strand_bias_pval_vaf_and_dp_parallel_each_indel(
  each_unique_indel_candidate,
  collapsed_read_counts,
  master_indel_record_table_no_dup_reads,
  mapq_threshold,
  bamPath
)
}
\arguments{
\item{each_unique_indel_candidate}{each_unique_indel_candidate}

\item{collapsed_read_counts}{collapsed_read_counts}

\item{master_indel_record_table_no_dup_reads}{master_indel_record_table_no_dup_reads}

\item{mapq_threshold}{mapq_threshold}

\item{bamPath}{bamPath}
}
\value{
row_to_add_to_cols_to_add
}
\description{
calculate_strand_bias_pval_vaf_and_dp_parallel_each_indel
}
