% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{convert_cigar_to_verbose_string}
\alias{convert_cigar_to_verbose_string}
\title{convert_cigar_to_verbose_string}
\usage{
convert_cigar_to_verbose_string(each_cigar_string)
}
\arguments{
\item{each_cigar_string}{each_cigar_string}
}
\value{
extended_string
}
\description{
convert_cigar_to_verbose_string
}
