% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{filter_and_annotate_calls}
\alias{filter_and_annotate_calls}
\title{filter_and_annotate_calls}
\usage{
filter_and_annotate_calls(
  master_indel_record_table,
  mapq_threshold,
  bamPath,
  number_cores,
  min_supporting_reads,
  min_vaf,
  min_read_depth,
  each_chromosome
)
}
\arguments{
\item{master_indel_record_table}{master_indel_record_table}

\item{mapq_threshold}{mapq_threshold}

\item{bamPath}{bamPath}

\item{number_cores}{number_cores}

\item{min_supporting_reads}{min_supporting_reads}

\item{min_vaf}{min_vaf}

\item{min_read_depth}{min_read_depth}

\item{each_chromosome}{each_chromosome}
}
\value{
collapsed_read_counts_with_strand_correction_pvalue_for_writing_filtered
}
\description{
filter_and_annotate_calls
}
