% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_primary_chroms}
\alias{get_primary_chroms}
\title{get_primary_chroms}
\usage{
get_primary_chroms(chr_in_bam)
}
\arguments{
\item{chr_in_bam}{chr_in_bam}
}
\value{
chr_in_bam
}
\description{
get_primary_chroms
}
