% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{getIndelCoverage}
\alias{getIndelCoverage}
\title{getIndelCoverage}
\usage{
getIndelCoverage(indel_chr, indel_start, indel_end, bamPath)
}
\arguments{
\item{indel_chr}{indel_chr}

\item{indel_start}{indel_start}

\item{indel_end}{indel_end}

\item{bamPath}{bamPath}
}
\value{
indel_read_depth
}
\description{
getIndelCoverage
}
