% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{getVaf}
\alias{getVaf}
\title{getVaf}
\usage{
getVaf(num_supporting_reads, indel_read_depth)
}
\arguments{
\item{num_supporting_reads}{num_supporting_reads}

\item{indel_read_depth}{indel_read_depth}
}
\value{
vaf
}
\description{
getVaf
}
