% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{refine_cigar_string}
\alias{refine_cigar_string}
\title{refine_cigar_string}
\usage{
refine_cigar_string(
  exploded_cigar_string,
  query_sequence_string,
  read_pos,
  query_read_length,
  each_chromosome,
  reference_sequence_test,
  number_leading_softclips
)
}
\arguments{
\item{exploded_cigar_string}{exploded_cigar_string}

\item{query_sequence_string}{query_sequence_string}

\item{read_pos}{read_pos}

\item{query_read_length}{query_read_length}

\item{each_chromosome}{each_chromosome}

\item{reference_sequence_test}{reference_sequence}

\item{number_leading_softclips}{number_leading_softclips}
}
\value{
exploded_cigar_string_for_refinement
}
\description{
refine_cigar_string
}
