% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{run_algo_all_reads_each_bam_region_with_simple_indel_padding}
\alias{run_algo_all_reads_each_bam_region_with_simple_indel_padding}
\title{run_algo_all_reads_each_bam_region_with_simple_indel_padding}
\usage{
run_algo_all_reads_each_bam_region_with_simple_indel_padding(
  row_num,
  per_bam_region_indel_records,
  mapq_threshold,
  bamPath,
  each_chromosome,
  sliding_windows_per_bam_region,
  verbose_arg,
  hg38_genome_chr_subset,
  flanking_region_length,
  min_indel_length
)
}
\arguments{
\item{row_num}{row_num}

\item{per_bam_region_indel_records}{per_bam_region_indel_records}

\item{mapq_threshold}{mapq_threshold}

\item{bamPath}{bamPath}

\item{each_chromosome}{each_chromosome}

\item{sliding_windows_per_bam_region}{sliding_windows_per_bam_region}

\item{verbose_arg}{verbose_arg}

\item{hg38_genome_chr_subset}{hg38_genome_chr_subset}

\item{flanking_region_length}{flanking_region_length}

\item{min_indel_length}{min_indel_length}
}
\value{
per_bam_region_indel_records
}
\description{
run_algo_all_reads_each_bam_region_with_simple_indel_padding
}
