% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{run_algo_on_one_read_explicit_args_with_simple_indel_padding}
\alias{run_algo_on_one_read_explicit_args_with_simple_indel_padding}
\title{run_algo_on_one_read_explicit_args_with_simple_indel_padding}
\usage{
run_algo_on_one_read_explicit_args_with_simple_indel_padding(
  per_bam_region_indel_records,
  refined_cigar_string,
  flanking_region_length,
  query_sequence_string,
  reference_sequence,
  read_pos,
  query_read_length,
  read_strand,
  read_name_record,
  number_leading_softclips,
  number_leading_hardclips,
  each_chromosome,
  min_indel_length
)
}
\arguments{
\item{per_bam_region_indel_records}{per_bam_region_indel_records}

\item{refined_cigar_string}{refined_cigar_string}

\item{flanking_region_length}{flanking_region_length}

\item{query_sequence_string}{query_sequence_string}

\item{reference_sequence}{reference_sequence}

\item{read_pos}{read_pos}

\item{query_read_length}{query_read_length}

\item{read_strand}{read_strand}

\item{read_name_record}{read_name_record}

\item{number_leading_softclips}{number_leading_softclips}

\item{number_leading_hardclips}{number_leading_hardclips}

\item{each_chromosome}{each_chromosome}

\item{min_indel_length}{min_indel_length}
}
\value{
per_bam_region_indel_records
}
\description{
run_algo_on_one_read_explicit_args_with_simple_indel_padding
}
