% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{run_indelfindr}
\alias{run_indelfindr}
\title{run_indelfindr}
\usage{
run_indelfindr(
  bamPath,
  bam_region_bin_size,
  verbose_arg,
  flanking_region_length,
  target_regions,
  number_cores,
  primary_chromosomes,
  min_indel_length,
  mapq_threshold,
  min_supporting_reads,
  min_vaf,
  min_read_depth,
  outname
)
}
\arguments{
\item{bamPath}{bamPath}

\item{bam_region_bin_size}{bam_region_bin_size}

\item{verbose_arg}{verbose_arg}

\item{flanking_region_length}{flanking_region_length}

\item{target_regions}{target_regions}

\item{number_cores}{number_cores}

\item{primary_chromosomes}{primary_chromosomes}

\item{min_indel_length}{min_indel_length}

\item{mapq_threshold}{mapq_threshold}

\item{min_supporting_reads}{min_supporting_reads}

\item{min_vaf}{min_vaf}

\item{min_read_depth}{min_read_depth}

\item{outname}{outname}
}
\description{
run_indelfindr
}
