% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{seq_with_uneven_last}
\alias{seq_with_uneven_last}
\title{Title}
\usage{
seq_with_uneven_last(from, to, by)
}
\arguments{
\item{from}{from}

\item{to}{to}

\item{by}{by}
}
\value{
vec to

to
}
\description{
Title
}
