% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{translate_cigar_index_to_ref_and_query_algo_dev_sliding}
\alias{translate_cigar_index_to_ref_and_query_algo_dev_sliding}
\title{translate_cigar_index_to_ref_and_query_algo_dev_sliding}
\usage{
translate_cigar_index_to_ref_and_query_algo_dev_sliding(
  sliding_window_per_read_start,
  sliding_window_per_read_end,
  reference_sequence_for_translation,
  query_sequence_string_for_translation,
  refined_cigar_string
)
}
\arguments{
\item{sliding_window_per_read_start}{sliding_window_per_read_start}

\item{sliding_window_per_read_end}{sliding_window_per_read_end}

\item{reference_sequence_for_translation}{reference_sequence_for_translation}

\item{query_sequence_string_for_translation}{query_sequence_string_for_translation}

\item{refined_cigar_string}{refined_cigar_string}
}
\value{
reference_allele_record alternate_allele_record cigar_string_record
}
\description{
translate_cigar_index_to_ref_and_query_algo_dev_sliding
}
