% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{translate_cigar_index_to_ref_and_query_v2}
\alias{translate_cigar_index_to_ref_and_query_v2}
\title{translate_cigar_index_to_ref_and_query_v2}
\usage{
translate_cigar_index_to_ref_and_query_v2(
  cigar_coords,
  cigar_coords_for_query,
  reference_sequence_for_translation,
  query_sequence_string_for_translation,
  refined_cigar_string
)
}
\arguments{
\item{cigar_coords}{cigar_coords}

\item{cigar_coords_for_query}{cigar_coords_for_query}

\item{reference_sequence_for_translation}{reference_sequence_for_translation}

\item{query_sequence_string_for_translation}{query_sequence_string_for_translation}

\item{refined_cigar_string}{refined_cigar_string}
}
\value{
list(reference_allele_record,alternate_allele_record,cigar_string_record)
}
\description{
translate_cigar_index_to_ref_and_query_v2
}
