% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{unexplode_cigar_string}
\alias{unexplode_cigar_string}
\title{unexplode_cigar_string}
\usage{
unexplode_cigar_string(exploded_cigar_string)
}
\arguments{
\item{exploded_cigar_string}{exploded_cigar_string}
}
\value{
cigar_record
}
\description{
unexplode_cigar_string
}
