% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{write.vcf}
\alias{write.vcf}
\title{write.vcf}
\usage{
write.vcf(master_indel_call_table, phased, output_filename)
}
\arguments{
\item{master_indel_call_table}{master_indel_call_table}

\item{phased}{phased}

\item{output_filename}{output_filename}
}
\description{
write.vcf
}
